+++ buildSrc/src/main/kotlin/Config.kt
@@ -1,6 +1,6 @@
 object Config {
 
     val TestDependencies: List<String> = listOf(
-        Libs.junit_jupiter_api, Libs.junit_jupiter_engine, Libs.rerunner_jupiter, Libs.kotlin_reflect, Libs.hamkrest
+        Libs.junit_jupiter_api, Libs.junit_jupiter_engine, Libs.kotlin_reflect, Libs.hamkrest
     )
 }