+++ http4k-cloudnative/src/test/kotlin/org/http4k/cloudnative/env/SecretTest.kt
@@ -4,8 +4,8 @@ import com.natpryce.hamkrest.allElements
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
 import com.natpryce.hamkrest.throws
-import org.junit.Test
 import org.junit.jupiter.api.Assertions.assertTrue
+import org.junit.jupiter.api.Test
 import java.util.concurrent.atomic.AtomicReference
 
 class SecretTest {