+++ http4k-cloudnative/src/test/kotlin/org/http4k/cloudnative/env/TimeoutTest.kt
@@ -2,7 +2,7 @@ package org.http4k.cloudnative.env
 
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.throws
-import org.junit.Test
+import org.junit.jupiter.api.Test
 import java.time.Duration
 
 class TimeoutTest {