+++ http4k-cloudnative/src/test/kotlin/org/http4k/filter/HandleUpstreamRequestFailedTest.kt
@@ -24,7 +24,7 @@ import org.http4k.core.Status.Companion.SERVICE_UNAVAILABLE
 import org.http4k.core.then
 import org.http4k.hamkrest.hasBody
 import org.http4k.hamkrest.hasStatus
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 class HandleUpstreamRequestFailedTest {